Updated UserService, linked api gateway to userservice

+++ storefront/src/main/java/com/fooddeliveryapp/storefront/exceptions/UserNotFoundException.java
@@ -2,7 +2,7 @@ package com.fooddeliveryapp.storefront.exceptions;
 
 public class UserNotFoundException extends RuntimeException {
 
-    public UserNotFoundException(Long id) {
+    public UserNotFoundException(int id) {
         super("User not found with id " + id);
     }
